api/requirements.txt
fastapi
uvicorn
httpx
pydantic
supabase
python-dotenv
numpy
rasterio
shapely
Pillow
pyproj
